Add in EP Serv call to submit handlers and abstract Feat Toggle to serv
