+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/PaperCaseCreateContestedAboutToSubmitHandler.java
@@ -15,6 +15,7 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.YesOrNo;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.CaseDataService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.CaseFlagsService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.IdamService;
+import uk.gov.hmcts.reform.finrem.caseorchestration.service.express.ExpressCaseService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.utils.refuge.RefugeWrapperUtils;
 
 import static uk.gov.hmcts.reform.finrem.caseorchestration.OrchestrationConstants.NO_VALUE;
@@ -29,15 +30,18 @@ public class PaperCaseCreateContestedAboutToSubmitHandler extends FinremCallback
     private final CaseFlagsService caseFlagsService;
     private final IdamService idamService;
     private final CaseDataService caseDataService;
+    private final ExpressCaseService expressCaseService;
 
     public PaperCaseCreateContestedAboutToSubmitHandler(FinremCaseDetailsMapper finremCaseDetailsMapper,
                                                         CaseFlagsService caseFlagsService,
                                                         IdamService idamService,
-                                                        CaseDataService caseDataService) {
+                                                        CaseDataService caseDataService,
+                                                        ExpressCaseService expressCaseService) {
         super(finremCaseDetailsMapper);
         this.caseFlagsService = caseFlagsService;
         this.idamService = idamService;
         this.caseDataService = caseDataService;
+        this.expressCaseService = expressCaseService;
     }
 
     @Override
@@ -92,6 +96,8 @@ public class PaperCaseCreateContestedAboutToSubmitHandler extends FinremCallback
         caseDataService.setFinancialRemediesCourtDetails(oldCaseDetails);
 
         caseData = finremCaseDetailsMapper.mapToFinremCaseData(oldCaseDetails.getData());
+
+        expressCaseService.setExpressCaseEnrollmentStatus(caseData);
 
         return GenericAboutToStartOrSubmitCallbackResponse.<FinremCaseData>builder().data(caseData).build();
     }
