+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/PaperCaseCreateContestedMidHandler.java
@@ -9,7 +9,6 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.model.EventType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseData;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseDetails;
-import uk.gov.hmcts.reform.finrem.caseorchestration.service.FeatureToggleService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.InternationalPostalService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.express.ExpressCaseService;
 
@@ -17,16 +16,13 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.service.express.ExpressCaseS
 @Service
 public class PaperCaseCreateContestedMidHandler extends FinremCallbackHandler {
 
-    private final FeatureToggleService featureToggleService;
     private final ExpressCaseService expressCaseService;
     private final InternationalPostalService postalService;
 
     public PaperCaseCreateContestedMidHandler(FinremCaseDetailsMapper finremCaseDetailsMapper,
-                                              FeatureToggleService featureToggleService,
                                               ExpressCaseService expressCaseService,
                                               InternationalPostalService postalService) {
         super(finremCaseDetailsMapper);
-        this.featureToggleService = featureToggleService;
         this.expressCaseService = expressCaseService;
         this.postalService = postalService;
     }
@@ -44,9 +40,7 @@ public class PaperCaseCreateContestedMidHandler extends FinremCallbackHandler {
         FinremCaseDetails caseDetails = callbackRequest.getCaseDetails();
         log.info("Invoking contested event {} mid event callback", EventType.NEW_PAPER_CASE);
 
-        if (featureToggleService.isExpressPilotEnabled()) {
-            expressCaseService.setExpressCaseEnrollmentStatus(caseDetails.getData());
-        }
+        expressCaseService.setExpressCaseEnrollmentStatus(caseDetails.getData());
 
         FinremCaseData caseData = caseDetails.getData();
         return GenericAboutToStartOrSubmitCallbackResponse.<FinremCaseData>builder()
