+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/solicitorcreatecase/SolicitorCreateContestedMidHandlerTest.java
@@ -12,7 +12,6 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.model.EventType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseData;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseDetails;
-import uk.gov.hmcts.reform.finrem.caseorchestration.service.FeatureToggleService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.InternationalPostalService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.SelectedCourtService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.express.ExpressCaseService;
@@ -21,7 +20,6 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.service.express.ExpressCaseS
 import java.util.Arrays;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.any;
-import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -42,8 +40,6 @@ class SolicitorCreateContestedMidHandlerTest {
     @Mock
     private SelectedCourtService selectedCourtService;
     @Mock
-    private FeatureToggleService featureToggleService;
-    @Mock
     private ExpressCaseService expressCaseService;
 
 
@@ -53,8 +49,7 @@ class SolicitorCreateContestedMidHandlerTest {
                 finremCaseDetailsMapper,
                 postalService,
                 selectedCourtService,
-                expressCaseService,
-                featureToggleService);
+                expressCaseService);
     }
 
     @Test
@@ -79,25 +74,13 @@ class SolicitorCreateContestedMidHandlerTest {
     }
 
     @Test
-    void testGivenExpressPilotEnabled_ThenExpressCaseServiceCalled() {
+    void testExpressCaseServiceCalled() {
         FinremCallbackRequest callbackRequest = buildFinremCallbackRequest();
         FinremCaseData caseData = callbackRequest.getCaseDetails().getData();
-        when(featureToggleService.isExpressPilotEnabled()).thenReturn(true);
 
         handler.handle(callbackRequest, AUTH_TOKEN);
 
         verify(expressCaseService).setExpressCaseEnrollmentStatus(caseData);
-    }
-
-    @Test
-    void testGivenExpressPilotDisabled_ThenExpressCaseServiceIsNotCalled() {
-        FinremCallbackRequest callbackRequest = buildFinremCallbackRequest();
-        FinremCaseData caseData = callbackRequest.getCaseDetails().getData();
-        when(featureToggleService.isExpressPilotEnabled()).thenReturn(false);
-
-        handler.handle(callbackRequest, AUTH_TOKEN);
-
-        verify(expressCaseService, never()).setExpressCaseEnrollmentStatus(caseData);
     }
 
     @Test
